+++ lib/features/settings/ui/library/library_settings_screen.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:go_router/go_router.dart';
+import 'package:storii/app/config/router.dart';
+import 'package:storii/app/init.dart';
+import 'package:storii/features/settings/ui/library/remember_sort_tile.dart';
+
+class LibrarySettingsTile extends ConsumerWidget {
+  const LibrarySettingsTile({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return ListTile(
+      trailing: const Icon(Icons.chevron_right),
+      leading: const Icon(Icons.library_books),
+      title: Text(l10n.librarySettings),
+      onTap: () {
+        context.push(AppRoute.librarySettings.path);
+      },
+    );
+  }
+}
+
+class LibrarySettingsScreen extends ConsumerWidget {
+  const LibrarySettingsScreen({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final textTheme = Theme.of(context).textTheme;
+
+    return Scaffold(
+      appBar: AppBar(
+        leading: IconButton(
+          onPressed: () => context.pop(),
+          icon: const Icon(Icons.arrow_back),
+        ),
+        title: Text(l10n.librarySettings, style: textTheme.titleLarge),
+      ),
+      body: const Column(children: [RememberSortTile()]),
+    );
+  }
+}